Deprecate RecurrencePattern
