+++ Ical.Net/CalendarComponents/RecurringComponent.cs
@@ -114,10 +114,10 @@ public abstract class RecurringComponent : UniqueComponent, IRecurringComponent
         set => Properties.Set("RRULE", value);
     }
 
-    public virtual RecurrencePattern? RecurrenceRule
+    public virtual RecurrenceRule? RecurrenceRule
     {
         get => RecurrenceRules?.FirstOrDefault();
-        set => RecurrenceRules = value is null ? [] : [value];
+        set => RecurrenceRules = value is null ? [] : [new(value)];
     }
 
     /// <summary>
